move images to separates folders
