+++ src/App/pages/LandingPage/Main/Carousel/images/left_arrow.png
