+++ src/App/pages/LandingPage/Main/Carousel/images/right_arrow.png
